Create TeamStats module, move team stat helpers into module
